move stale interval to heartbeat file
